WorkInProgress: Beginning of making drag&drop of assets and sorting of slides and tray work. Add 2 css files two show templates in their correct arrangement. Show the bigger preview image in slots to improve the impression of the slide in the editor.

+++ app/controllers/tray_positions_controller.rb
@@ -8,16 +8,18 @@ class TrayPositionsController < ApplicationController
   end
   def replace
     respond_to do |format|
-      format.html do
-        current_user.tray_positions.find(params[:del]).each do |tp|
-          if !tp.clipboard_type  && (tp.asset.classifications_count || 0) == 0
-            tp.asset.content.destroy
-          else
-            tp.destroy
+      if params[:del]
+        format.html do
+          current_user.tray_positions.find(params[:del]).each do |tp|
+            if !tp.clipboard_type  && (tp.asset.classifications_count || 0) == 0
+              tp.asset.content.destroy
+            else
+              tp.destroy
+            end
           end
+          flash[:notice] = "Tray has been updated!"
+          redirect_to :back
         end
-        flash[:notice] = "Tray has been updated!"
-        redirect_to :back
       end
       format.json do
         new_tp = nil
